requirements.txt
Adafruit-ADS1x15==1.0.2
Adafruit-Blinka==8.25.0
adafruit-circuitpython-ads1x15==2.2.24
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-requests==2.0.2
adafruit-circuitpython-typing==1.9.5
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.54.0
Adafruit-PureIO==1.1.11
av==11.0.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
et-xmlfile==1.1.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
mariadb==1.1.8
numpy==1.26.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
picamera2==0.3.14
pidng==4.0.9
piexif==1.1.3
Pillow==10.1.0
protobuf==4.25.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyftdi==0.55.0
pyparsing==3.1.1
pyserial==3.5
python-prctl==1.8.1
pyusb==1.2.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simplejpeg==1.7.2
spidev==3.6
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
v4l2-python3==0.3.2
